requirements.txt
# Core Dependencies
pyspark==3.5.0
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Validation
great-expectations==0.18.5
pydantic==2.5.2

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
click==8.1.7

# Azure (Opcional - apenas para upload manual ou testes)
# Descomente se precisar fazer upload manual de arquivos para Azure Storage
# azure-storage-blob==12.19.0
# azure-identity==1.15.0

# Database Drivers (Opcional - apenas se usar exportação SQL)
# Spark geralmente baixa automaticamente via spark.jars.packages
# Mas você pode instalar localmente se necessário:
# pyodbc==5.0.1  # Para SQL Server (Windows)
# psycopg2-binary==2.9.9  # Para PostgreSQL
# pymysql==1.1.0  # Para MySQL

